Allow numeric arguments to be numbers.
The speed, runlevels, bootloader_timeout and logout_timeout
parameters are validated as strings of digits, so they cannot
be specified as numbers. This means that in Hiera, you'd have to
write:

    serial_console::speed: '9600'

Rather than:

    serial_console::speed: 9600

By putting the argument to validate_re in double quotes, we
force a string context on numbers, so validation succeeds on
both numbers and strings of digits (i.e. both 9600 and '9600').